third_party/blink/web_tests/fast/forms/range/input-range-validation-message-expected.txt

Test for validationMessage IDL attribute for <input type=range>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

No message
PASS testIt("", "", "") is ""
Value missing
PASS testIt("", "", "") is ""
Type mismatch
PASS testIt("foo", "", "") is ""
Range overflow
PASS testIt("200", "", "100") is ""
Range underflow
PASS testIt("10", "50", "100") is ""
Step mismatch
PASS testIt("55", "0", "100", "10") is ""
Range invalid
PASS testIt("10", "100", "50") is ""
PASS successfullyParsed is true

TEST COMPLETE
